Convert an OpenML task to mlr.

Converts an OMLTask to a list of Task, ResampleInstance and Measure.

convertOMLTaskToMlr(obj, measures = NULL,
  mlr.task.id = "<oml.data.name>", ignore.flagged.attributes = TRUE,
  drop.levels = TRUE, verbosity = NULL)

Arguments

obj
[OMLTask]
The OML task object that should be converted.

measures
[Measure]
Additional measures that should be computed.

mlr.task.id
[character(1)]
Id string for Task object. The strings <oml.data.name>, <oml.data.id>, <oml.data.version> and <oml.task.id> will be replaced by their respective values contained in the OMLTask object. Default is <oml.data.name>.

ignore.flagged.attributes
[logical(1)]
Should those features that are listed in the data set description slot “ignore.attribute” be removed? Default is TRUE.

drop.levels
[logical(1)]
Should empty factor levels be dropped in the data? Default is TRUE.

verbosity
[integer(1)]
Print verbose output on console? Possible values are:
0: normal output,
1: info output,
2: debug output.
Default is set via setOMLConfig.

Value

[list] A list with the following objects:

mlr.task

[Task]

mlr.rin

[ResampleInstance]

mlr.measures

[list of Measures to optimize for.

See also

Examples

# \dontrun{
# 	library("mlr")
# 	vinnieOML = getOMLTask(task.id = 4845)
# 	vinnieMlr = convertOMLTaskToMlr(vinnieOML)
# }