List all registered OpenML flows.

The returned data.frame contains the flow id “fid”, the flow name (“full.name” and “name”), version information (“version” and “external.version”) and the uploader (“uploader”) of all registered OpenML flows.

listOMLFlows(tag = NULL, limit = NULL, offset = NULL,
  verbosity = NULL)

Arguments

tag
[character]
If not NULL only entries with the corresponding tags are listed.

limit
[numeric(1)]
Optional. The maximum number of entries to return. Without specifying offset, it returns the first 'limit' entries. Setting limit = NULL returns all available entries.

offset
[numeric(1)]
Optional. The offset to start from. Should be indices starting from 0, which do not refer to IDs. Is ignored when no limit is given.

verbosity
[integer(1)]
Print verbose output on console? Possible values are:
0: normal output,
1: info output,
2: debug output.
Default is set via setOMLConfig.

Value

[data.frame].

Note

This function is memoised. I.e., if you call this function twice in a running R session, the first call will query the server and store the results in memory while the second and all subsequent calls will return the cached results from the first call. You can reset the cache by calling forget on the function manually.

See also

Examples

# \dontrun{
#   flows = listOMLFlows()
#   tail(flows)
# }