List the first 5000 OpenML tasks.

The returned data.frame contains the task_id, the data set id data.id, the status and some describing data qualities. Note that by default only the first 5000 data sets will be returned (due to the argument “limit = 5000”).

listOMLTasks(task.type = NULL, estimation.procedure = NULL,
  evaluation.measures = NULL, number.of.instances = NULL,
  number.of.features = NULL, number.of.classes = NULL,
  number.of.missing.values = NULL, tag = NULL, data.name = NULL,
  data.tag = NULL, limit = 5000, offset = NULL, status = "active",
  verbosity = NULL)

Arguments

task.type
[character(1)]
If not NULL, only tasks belonging to the given task type are listed. Use listOMLTaskTypes()$name to see possible values for task.type. The default is NULL, which means that tasks with all available task types are listed.

estimation.procedure
[character]
If not NULL, only tasks belonging the given estimation procedures are listed. Use listOMLEstimationProcedures()$name to see possible values for estimation.procedure. The default is NULL, which means that tasks with all available estimation procedures are listed.

evaluation.measures
[character]
If not NULL, only tasks belonging the given evaluation measures are listed. Use listOMLEvaluationMeasures()$name to see possible values for evaluation.measures. The default is NULL, which means that tasks with all available evaluation measures are listed.

number.of.instances
[numeric(1) | numeric(2)]
If not NULL, subsets the entries with respect to the given values or, if a vector of length 2 is passed, the given ranges.

number.of.features
[numeric(1) | numeric(2)]
If not NULL, it subsets the entries with respect to the given values or, if a vector of length 2 is passed, the given range.

number.of.classes
[numeric(1) | numeric(2)]
If not NULL, subsets the entries with respect to the given values or, if a vector of length 2 is passed, the given ranges.

number.of.missing.values
[numeric(1) | numeric(2)]
If not NULL, subsets the entries with respect to the given values or, if a vector of length 2 is passed, the given ranges.

tag
[character]
If not NULL only entries with the corresponding tags are listed.

data.name
[character(1)]
Name of the data set.

data.tag
[character(1)]
Refers to the tag of the dataset the task is based on. If not NULL only tasks with the corresponding data.tag are listed.

limit
[numeric(1)]
Optional. The maximum number of entries to return. Without specifying offset, it returns the first 'limit' entries. Setting limit = NULL returns all available entries.

offset
[numeric(1)]
Optional. The offset to start from. Should be indices starting from 0, which do not refer to IDs. Is ignored when no limit is given.

status
[character]
Subsets the results according to the status. Possible values are {"active", "deactivated", "in_preparation", "all"}. Default is "active".

verbosity
[integer(1)]
Print verbose output on console? Possible values are:
0: normal output,
1: info output,
2: debug output.
Default is set via setOMLConfig.

Value

[data.frame].

Note

This function is memoised. I.e., if you call this function twice in a running R session, the first call will query the server and store the results in memory while the second and all subsequent calls will return the cached results from the first call. You can reset the cache by calling forget on the function manually.

See also

Examples

# \dontrun{
#   tasks = listOMLTasks()
#   head(tasks)
# }