Construct OMLRun.

More details about the elements of a OMLRun can be found in the XSD scheme.

makeOMLRun(run.id = NA_integer_, uploader = NA_integer_,
  uploader.name = NA_character_, task.id, task.type = NA_character_,
  task.evaluation.measure = NA_character_, flow.id = NA_integer_,
  flow.name = NA_character_, setup.id = NA_integer_,
  setup.string = NA_character_, error.message = NA_character_,
  parameter.setting = list(), tags = NA_character_,
  predictions = NULL, input.data = makeOMLIOData(),
  output.data = makeOMLIOData())

Arguments

run.id
[numeric(1)]
ID of the run. Added by server. Ignored when uploading a run.

uploader
[numeric(1)]
ID of the user that uploaded the run. Added by server. Ignored when uploading a run.

uploader.name
[character(1)]
Name of the user that uploaded the run. Ignored when uploading a run.

task.id
[numeric(1)]
ID of the task that is solved in this run. This ID is given in the task description.

task.type
[character(1)]
Task type of the run. See listOMLTaskTypes for all possible types.

task.evaluation.measure
[character(1)]
Evaluation measure used in the run.

flow.id
[character(1)]
ID of the flow used to solve the task. Returned by the API when you upload the flow, or given in the flow description when you download an existing flow.

flow.name
[character(1)]
Name of the flow.

setup.id
[numeric(1)]
Unique ID of the used setup. Ignored when uploading a run (i.e., it will be searched based on the parameter settings).

setup.string
[character(1)]
The CLI string that can invoke the learner with the correct parameter settings. This argument is optional.

error.message
[character(1)]
Whenever an error occurs during the run, this can be reported here.

parameter.setting
[list]
A list of OMLRunParameters containing information on the parameter settings.

tags
[character]
Optional tags describing the run.

predictions
[data.frame]
The predictions of the run.

input.data
[OMLIOData]
All data that served as input for the run. Added by server. Ignored when uploading.

output.data
[OMLIOData]
All data that was the output of this run, i.e., predictions, evaluation scores. Most of this will be added by the server, but users can also provide evaluation scores for their own evaluation measures.

See also