Reproduce the Run

Uses the ID of the run and tries to reproduce its results by downloading the flow and applying it to the respective task.

runTaskFlow(task, flow, par.list, seed = 1, predict.type = NULL,
  verbosity = NULL, models = TRUE)

Arguments

task
[OMLTask]
An OpenML task.

flow
[OMLFlow]
Flow that is applied to the Task.

par.list
[list|OMLRunParList]
Can be either a named list containing the hyperparameter values or a OMLRunParList.

seed
[numeric(1)|OMLSeedParList ]
Set a seed to make the run reproducible. Default is 1 and sets the seed using set.seed(1).

predict.type
[character(1)]
Optional. See setPredictType. Default is "response".

verbosity
[integer(1)]
Print verbose output on console? Possible values are:
0: normal output,
1: info output,
2: debug output.
Default is set via setOMLConfig.

models
[logical(1)]
This argument is passed to benchmark. Should all fitted models be stored in the ResampleResult? Default is TRUE.

Value

[OMLMlrRun], an OMLRun.