Run mlr learner on OpenML task.

Run task with a specified learner from mlr and produce predictions. By default, the evaluation measure contained in the task is used.

runTaskMlr(task, learner, measures = NULL, verbosity = NULL,
  seed = 1, scimark.vector = NULL, models = TRUE, ...)

Arguments

task
[OMLTask]
An OpenML task.

learner
[Learner]
Learner from package mlr to run the task.

measures
[Measure]
Additional measures that should be computed.

verbosity
[integer(1)]
Print verbose output on console? Possible values are:
0: normal output,
1: info output,
2: debug output.
Default is set via setOMLConfig.

seed
[numeric(1)|OMLSeedParList ]
Set a seed to make the run reproducible. Default is 1 and sets the seed using set.seed(1).

scimark.vector
[numeric(6)]
Optional vector of performance measurements computed by the scientific SciMark benchmark. May be computed using the rscimark R package. Default is NULL, which means no performance measurements.

models
[logical(1)]
This argument is passed to benchmark. Should all fitted models be stored in the ResampleResult? Default is TRUE.

...
[any]
Further arguments that are passed to convertOMLTaskToMlr.

Value

[list] Named list with the following components:

run

The OMLRun object.

bmr

Benchmark result returned by benchmark.

flow

The generated OMLFlow object.

See also

Examples

# \dontrun{
#   library(mlr)
#   ## run a single flow (learner) on a single task
#   task = getOMLTask(57)
#   lrn = makeLearner("classif.rpart")
#   res = runTaskMlr(task, lrn)
#   ## the result "res" is a list, storing information on the actual "run", the
#   ## corresponding benchmark result "bmr" and the applied "flow"
# }