Saves a list of OpenML configuration settings to file.

The new configuration is automatically assigned via setOMLConfig if all checks pass. If you don't set a certain option, package defaults will be inserted into the file.

saveOMLConfig(server = NULL, verbosity = NULL, apikey = NULL,
  cachedir = NULL, arff.reader = NULL, confirm.upload = NULL,
  overwrite = FALSE)

Arguments

server
[character(1)]
URL of the XML API endpoint.

verbosity
[integer(1)]
Verbosity level. Possible values are 0 (normal output), 1 (info output), 2 (debug output).

apikey
[character(1)]
Your OpenML API key. Log in to OpenML, move to your profile to get it.

cachedir
[character(1)]
Path to the cache directory.

arff.reader
[character(1)]
Name of the package which should be used to parse arff files. Possible are “RWeka”, which is the default and “farff”.

confirm.upload
[logical(1)]
Should the user be asked for confirmation before upload of OML objects?

overwrite
[logical(1)]
Should an existing file be overwritten? Default is FALSE.

See also