Settter for configuration settings.

Set and overwrite configuration settings.

setOMLConfig(server = NULL, verbosity = NULL, apikey = NULL,
  cachedir = NULL, arff.reader = NULL, confirm.upload = NULL)

Arguments

server
[character(1)]
URL of the XML API endpoint.

verbosity
[integer(1)]
Verbosity level. Possible values are 0 (normal output), 1 (info output), 2 (debug output).

apikey
[character(1)]
Your OpenML API key. Log in to OpenML, move to your profile to get it.

cachedir
[character(1)]
Path to the cache directory.

arff.reader
[character(1)]
Name of the package which should be used to parse arff files. Possible are “RWeka”, which is the default and “farff”.

confirm.upload
[logical(1)]
Should the user be asked for confirmation before upload of OML objects?

Value

Invisibly returns a list of configuration settings.

See also