Upload an OpenML run.

Share a run of a flow on a given OpenML task by uploading it to the OpenML server.

uploadOMLRun(run, upload.bmr = FALSE, tags = NULL,
  confirm.upload = NULL, verbosity = NULL, ...)

Arguments

run
[OMLRun|OMLMlrRun]
The run that should be uploaded. Either a OMLRun or a run created with OMLMlrRun.

upload.bmr
[logical(1)]
Should the Benchmark result created by benchmark function be uploaded? If set to TRUE and the flow is created via makeTuneWrapper, an arff file that contains the hyperparameter optimization trace is also uploaded.

tags
[character]
The tags that should be added after uploading.

confirm.upload
[logical(1)]
Should the user be asked to confirm the upload? Default is the setting from your config.

verbosity
[integer(1)]
Print verbose output on console? Possible values are:
0: normal output,
1: info output,
2: debug output.
Default is set via setOMLConfig.

...
Not used.

Value

[invisible(numeric(1))]. The run ID.

Note

This function will reset the cache of listOMLRuns and listOMLRunEvaluations on success.

By default you will be asked to confirm the upload. You can deactivate the need for confirmation by setting “confirm.upload = TRUE” via setOMLConfig or set the corresponding argument each time you call the function.

See also