Upload a task to the OpenML server.

Share a task by uploading it to the OpenML server.

uploadOMLTask(task.type, data.id, target.feature, estimation.procedure,
  evaluation.measure = NULL, tags = NULL, description = NULL,
  confirm.upload = NULL, verbosity = NULL)

Arguments

task.type
[character(1)]
The type of the task to upload. See listOMLTaskTypes() to list all valid task types.

data.id
[integer(1)]
ID of the data set.

target.feature
[character(1)]
The target feature of the dataset.

estimation.procedure
[character(1)]
The estimation procedure for the evaluation. See listOMLEstimationProcedures() to list all procedures.

evaluation.measure
[character(1)]
The evaluation measure for the evaluation. See listOMLEvaluationMeasures() to list all possible measures.

tags
[character]
The tags that should be added after uploading.

description
[character(1)|OMLDataSetDescription]
Either an OMLDataSetDescription or a character(1) that describes the data. For the latter, all other relevant information is autogenerated from the Task.

confirm.upload
[logical(1)]
Should the user be asked to confirm the upload? Default is the setting from your config.

verbosity
[integer(1)]
Print verbose output on console? Possible values are:
0: normal output,
1: info output,
2: debug output.
Default is set via setOMLConfig.