Construct OMLFlow.

More details about the elements of a OMLFlow can be found in the XSD scheme.

makeOMLFlow(flow.id = NA_integer_, uploader = NA_integer_, name,
  version = NA_character_, external.version = NA_character_,
  description, creator = NA_character_, contributor = NA_character_,
  upload.date = NA_character_, licence = NA_character_,
  language = "English", full.description = NA_character_,
  installation.notes = NA_character_, dependencies = NA_character_,
  bibliographical.reference = NULL, implements = NA_character_,
  parameters = NULL, components = NULL, qualities = NULL,
  tags = NA_character_, source.url = NA_character_,
  binary.url = NA_character_, source.format = NA_character_,
  binary.format = NA_character_, source.md5 = NA_character_,
  binary.md5 = NA_character_, source.path = NA_character_,
  binary.path = NA_character_, object = NULL)

Arguments

flow.id
[integer(1)]
ID of the flow. Generated by the server, based on name and version of the flow. Ignored when uploaded manually.

uploader
[integer(1)]
The user that uploaded the flow. Added by the server. Ignored when uploaded manually.

name
[character(1)]
The name of the flow. Name-version combinations should be unique.
Allowed characters: () [] a-z A-Z 0-9 . _ - +

version
[character(1)]
The version of the flow. Default is 1.0. Ignored at upload time.

external.version
[character(1)]
An external version, defined by the user. In combination with the name, it must be unique.

description
[character(1)]
A user description of the flow.

creator
[character]
Optional. The persons/institutions that created the flow.

contributor
[character]
Optional. (Minor) contributors to the workflow

upload.date
[character(1)]
The date on which the flow was uploaded.
Format YYYY-mm-ddThh:MM:SS. Added by the server. Ignored when uploaded manually.

licence
[character(1)]
Optional. Default is none, meaning Public Domain or "don't know/care".

language
[character(1)]
Optional. Starts with one upper case letter, rest is lower case. Default is English.

full.description
[character(1)]
Optional. Full description of the workflow, e.g, man pages filled in by tool. This is a much more elaborate description than given in the 'description field'. It may include information about all components of the workflow.

installation.notes
[character(1)]
Optional. Additional hints on how to run the flow.

dependencies
[character(1)]
Optional. The dependencies of the flow.

bibliographical.reference
[list]
An optional list containing information on bibliographical references in form of OMLBibRef.

implements
[character(1)]
Ontological reference.

parameters
[list]
The parameters of the flow. A list containing OMLFlowParameters.

components
[list]
A list containing OMLFlows. Typically components of a workflow or subfunctions of an algorithm (e.g. kernels). Components can have their own parameters.

qualities
[list]
Qualities of the algorithm. Each member of the list is an OMLFlowQuality.

tags
[character]
Tags describing the algorithm.

source.url
[character(1)]
URL from which the source code can be downloaded. Added by the server. Ignored when uploaded manually.

binary.url
[character(1)]
URL from which the binary can be downloaded. Added by the server. Ignored when uploaded manually.

source.format
[character(1)]
Format of the source file.

binary.format
[character(1)]
Format of the binary file.

source.md5
[character(1)]
MD5 checksum to check if the source code was uploaded correctly.

binary.md5
[character(1)]
MD5 checksum to check if the binary code was uploaded correctly.

source.path
[character(1)]
The path to the cached source file, once getOMLFlow was run.

binary.path
[character(1)]
The path to the cached binary file, once getOMLFlow was run.

object
[any]
(optional) Any R object referring to the flow.

See also